Move delay under utils
